First function almost done

+++ app/src/main/java/com/riseup/riseup_users/model/UserModel.kt
@@ -6,7 +6,7 @@ import kotlin.collections.ArrayList
 data class UserModel (
     var cel:String ="",
     var email:String = "",
-    var diamonds:Int = 0,
+    var diamonds:Double = 0.0,
     var age:Int = 0,
     var favorites:ArrayList<String>? = null,
     val id:String = "",
